requirements.txt
hydra-core>=1.3.2
Pillow>=10.0.0
reportlab>=4.0.4
omegaconf>=2.3.0
PyYAML>=6.0.1
